mkimage: fit: add support to encrypt image with aes
This commit add the support of encrypting image with aes
in mkimage. To enable the ciphering, a node cipher with
a reference to a key and IV (Initialization Vector) must
be added to the its file. Then mkimage add the encrypted
image to the FIT and add the key and IV to the u-boot
device tree.

Change-Id: I0ec646905ceb64e64878f6c670f28b2c01fc7473
Signed-off-by: Philippe Reynes <philippe.reynes@softathome.com>